[post-order] Add an API to PostOrderFunctionInfo that maps a BB to its RPO and PO number.
We often times create this map in passes so that we can map from a BB -> ID ->
StateObject. Since it seems a fundamental part of the API, lets just stick it on PostOrderFunctionInfo.

I also moved a few declarations into PostOrderFunctionInfo. Calling getPostOrder
on the analysis instead of the function info is an artifact from an earlier
iteration of PostOrderAnalysis when it was not based on function info base. We
should always get the function info to follow the normal function info base
pattern.

I rewrote the PostOrderAnalysis APIs to use the new PostOrderFunctionInfo APIs.
I am going to move them over piece by piece.

Swift SVN r31663