Camel casing: cope with acronym/prefix conflicts when matching names to types.
When matching a word from a name to a word from a type, handle the
ambiguity between type prefixes (NS, UI) and acronyms (URL, HTTP) by
matching the name word to the end of the type word so long as we don't
have lowercase letters or underscores preceding the match. This
allows a name word "URL" or "HTTP" to match a type word like "NSURL"
or "NSHTTP" without having to hardcode knowledge of prefixes.

This mostly affects the omit-needless-words mode, but can also help us
identify more factory methods that can become initializers.

Swift SVN r31703