add a somewhat gross hack to CSDiag's to tell it to not retypecheck OverloadedDeclRefExpr or OverloadedMemberRefExpr when there is no contextual type information available.  The problem is that CSRanking will take a look at the various solutions formed by picking each member of the set, and will arbitrarily rank them against each other based on how specific the candidates are.  The problem with this is that the constraints on the candidates are being resolved by UnresolvedType, which means that we end up accidentally pruning the overload set too early.
This can lead to incorrect diagnostics that *should* have been ambiguity
diagnostics, such as the example in TypeCoercion/overload_noncall.swift.
It also is causing me other grief as I'm trying to make the call analysis
diagnostics more specific and the lack of the proper candidates is 
triggering badness.

The actual change to the testsuite here is minor, but not all good.  It will
be re-won by later changes.



Swift SVN r31744