Allow inout types in foreign function signatures, NFC
Previously we would ignore inout when bridging, but now we need
to take it into account for Clang-generated accessors.

Note that @block_storage must be special cased, because we always
require @inout to be specified together with @block_storage, and
@inout @block_storage is not a pointer type to some underlying
type, but rather a special block pointer type that comes directly
from the Clang AST context.

Swift SVN r31778