[sil-combine] Re-factor the implementation of propagateConcreteTypeOfInitExistential.
Aggregate all the relevant logic into propagateConcreteTypeOfInitExistential and its helper functions.
Support propagation of the concrete type information through copy_addr instructions.

This improves the performance of the test-case similar to the one reported in rdar://22569911, but with the locally defined existential.

The logic of the type propagation will soon be moved/implemented in the dedicated analysis pass (rdar://22613131)

Swift SVN r31796