Minor refactoring of call graph construction.
Directly build call graph edges rather than building a callee set and
then returning that to use to build a call graph edge, since the
constructor we want to call is based on whether we're building a
singleton set or not and we know that when we construct the set.

Once we handle cases with multiple callees, we'll again call into
helpers to build those sets, but those sets will likely be precomputed,
or computed on-demand and cached, and then we'll call into the
appropriate constructor for CallGraphEdge to make the edge for that
non-singleton set.

Swift SVN r31829