Create a call graph edge for every FullApplySite.
Rather than bailing in cases where we do not currently know what is
going to be called, create an edge with the callee set to nullptr.

The interface for getting the edge for a call site will return an empty
set for these.

Grows the allocated memory for the call graph by 2% when compiling the
stdlib.

Doing this makes it possible to walk all the FullApplySites in a
function via the call graph rather than walking the body of the
function. It also removes the need for clients to special case the "no
edge found" case when querying the call graph (instead they get an edge
that is known to not have all callees represented).

Swift SVN r31851