ASTPrinter: when skipping parameter names, indicate argument labels when they aren't the default.
For a function like

  func foo(string aString: String) { ... }

we previously printed

  func foo(string: String)

under -skip-parameter-names. Instead, print

  func foo(string _: String)

to make it clear that "string" is an argument label here.

The only thing that seems to use this flag are the omit-needless-words
experiments, and this change is tuned to minimally change the output
of such experiments for diff'ability.

Swift SVN r31907