Add support to the call graph for class_method.
We will now create edges for applies of class_method, but do not yet
determine whether the potential callees are comprised of the complete
set of all possible callees.

We don't currently have a pass that executes just for the purpose of
printing the results of call graph construction, so that will come next,
along with the start of some tests based on using that pass.

Swift SVN r31957