Don't store root nodes in the call graph.
The vector of root nodes is hard to maintain as soon as we remove nodes from the graph.
The root nodes are currently only used during SCC building. And here we can simpliy iterate over all functions
(which we do anyway, so there is no increase of complexity)



Swift SVN r32069