[let-properties-opts] Don't remove assignments to let properties from initializers (part 2).
Let properties optimizations were removing assignments to let properties (inside initializers), which are private or internal and compiled with WMO. It seemed to be safe, because all explicit reads of those properties were replaced by their compile-time constant values. But it turned out to be problematic if one uses Mirrors, because in this case, new values of a given type are constructed implicitly by dynamically invoking an initializer through the introspection.

This fixes rdar://22725946

Swift SVN r32178