[test] Add a test for weak properties in 'let' structs.
A weak property in a struct can become nil even if the struct is immutable,
kind of as if the weak reference were actually indirected through a
(mutable) class. We might decide to disallow this someday, or formalize it
better, but as long as we allow it we shouldn't misoptimize it.

Swift SVN r32248