Temporarily disable the SLPVectorizer
The SLPVectorizer vectorizes loads from a packed struct of i2s.

 {<i2, i2, i2, i2>}

The problem with that is that LLVM codegens scalar accesses to such a type as
if it where a packed i8 struct.
So when we emit a vector load of <4 x i2> (i.e one byte) and extract any non
zero index element we get garbage. Disable the SLPVectorizer until it is fixed.

rdar://22673382

Swift SVN r32299