Refactor out apply related combines from SILCombinerVisitors.cpp => SILCombinerApplyVisitors.cpp.
This includes code related to optimizing partial_apply, apply, and try_apply.

Original SILCombinerVisitors was relatively large with ~3300 lines of code. This
is a part of a series of commits that slims down the file. This commit reduces
SILCombinerVisitors to ~1050 lines of code, a 70% reduction.

Now the code remaining in SILCombinerVisitors only deals with some miscellaneous
canonicalizations related to:

1. Loads.
2. Ref Counting instructions.
3. Existential Initializations.
4. Projections/Aggregates.
5. Control Flow Canonicalizations.

I am going to rename SILCombinerVisitors.cpp to SILCombinerMiscVisitors.cpp. If
it gets too big, we can split it again.

Swift SVN r32327