CMake: fix separate overlay builds
When building overlay separately, look for dylibs in the overlay
build directory first, and only then fall back to the compiler resource
directory.

If the overlay changes the ABI, we shouldn't even try the dylibs in the
compiler resource directory, that is certain to fail the build or
result in miscompiles.

rdar://22866170

Swift SVN r32382