SILGen: Fix miscompile if self accessed multiple times while forming delegating init call
When emitting a RebindSelfInConstructorExpr, the first access of
'self' would take the self value out of the box, since the call
consumes it. This caused a miscompile if we have to load instance
variables to form the call, since then we would try to load from
the box again, dereferencing a null pointer. Now, store the taken
self value and borrow it on subsequent access.

This is a regression from r31141.

Fixes <rdar://problem/22939666>.

Swift SVN r32407