Minor refactoring of call graph interface.
We only need to know whether a callee set for a class_method is complete
when we go to create the call graph edge, so defer that determination
until that point. This is currently just hard coded to false, but a
subsequent commit will actually make that determination.

Swift SVN r32463