Change DI to calculate if 'self' was consumed by a delegating init call
This is done by adding a second variable to the dataflow analysis,
SelfConsumed. This is set to No in the block introducing the self
value and Yes in any error blocks of a try_apply call of an
initializer on self.

Previously we would model self as having been fully initialized at
this point, storing a null pointer into the box. This relied on the
fact that releasing a null pointer is a no-op, but it causes some
difficulties with designated init early returns, so we have to
handle it with more rigor now.

NFC for now, since nothing uses the result of the analysis yet.

Swift SVN r32531