SimplifyCFG: fix a crash in the try_apply(convert_function) to apply conversion.
The convert_function can convert to a function type with different argument types.
This was not checked.
I tried to keep the fix as simple as possible and just bail out if there is a mismatch.
Eventually we should allow the replacement and insert casts as we already do for the return type.

rdar://problem/23037566




Swift SVN r32547