Iterative type checker: thread an UnsatisfiedDependency through type resolution.
Start threading an (optional) UnsatisfiedDependency callback through type resolution, poke it when we're going to perform name lookup on a particular name.

As part of this, teach type resolution to distinguish between "early return due to unsatisfied dependencies" and "early return due to error", and not to mark type representations as invalid while failing.

This is all inactive code.

Swift SVN r32567